base completed
I am able to input a scripture reference, with a translation, and get a verse from the api. I then store that information into a tinydb which is just a json file. I will eventually use this json file for a bunch of things, but for now i just needed to create it.

I need to add error checks everywhere because someone might misspell a book name in the translation that they are wanting, or they might add a book number or verse number that does not exist in the approriate context