Can now insert scripture with little to no errors
There are checks for Translation, Book Name, Chapter Number, and Verse Number. All of them have to be real before we even try to search for the verse.

The user can insert scripture into the database so far, but that's it. Maybe I should confirm if the verse that the person entered is actually what they want