Multiple error messages for scripture reference
When a user inputs something like 'Jon 99:70 ABC' they are not told that the translation isn't right, the book name isn't right, the chapter number isn't right, and the verse number isn't right. This is so that the user can see what exactly went wrong with their reference. Obviously, if the book name is not right, then the chapter and verse are not going to be right