Solved long book name issue
Some translation of the bible have book titles that are long. Romans in NKJV is "The Epistle of Paul the Apostle to the ROMANS". So to solve this, we take the user input of hopefully "Romans", in this example, and try to find it in NKJV bible translation. When we cannot find it, we try to find Romans in the WEB bible translation. If we find it in WEB, we remember the book id and use that id in the NKJV bible translation to grab the verse we want.

This might not solve all the issue because we are assuming that the book ids are the same throughout translations. I will attempt to tackle that problem if I ever run into it